refactor: добавить пакет с классами RowMapper, внести изменения в код классов с использованием RowMapper

+++ src/main/java/ru/yandex/practicum/filmorate/model/Genre.java
@@ -6,15 +6,15 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Genre {
-    private int id;
+    private long id;
     private String name;
 
-    public Genre(int id, String name) {
+    public Genre(long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Genre(int id) {
+    public Genre(long id) {
         this.id = id;
     }
 }
